+++ resources/compiled.csv
@@ -1,0 +1,19 @@
+"MatchId","TeamId","Color","AutoLowCone","AutoLowCube","AutoMidCone","AutoMidCube","AutoHighCone","AutoHighCube","AutoLeftComm","AutoDocked","AutoEngaged","TeleLowCone","TeleLowCube","TeleMidCone","TeleMidCube","TeleHighCone","TeleHighCube","TeleLeftComm","TeleDocked","TeleEngaged","TeleTeamRole","TeleDirtyPlay","EndGameNotes","Won"
+"0","1810","blue","0","0","0","0","0","0","0","yes","no","0","0","0","0","0","0","0","yes","no","runner","no","","yes"
+"1","1811","blue","1","1","1","1","1","1","1","yes","no","1","1","1","1","1","1","1","yes","no","runner","no","","yes"
+"2","1812","blue","2","2","2","2","2","2","2","yes","no","2","2","2","2","2","2","2","yes","no","runner","no","","yes"
+"3","1810","blue","0","0","0","0","0","0","0","yes","no","0","0","0","0","0","0","0","yes","no","runner","no","","yes"
+"4","1811","blue","1","1","1","1","1","1","1","yes","no","1","1","1","1","1","1","1","yes","no","runner","no","","yes"
+"5","1812","blue","2","2","2","2","2","2","2","yes","no","2","2","2","2","2","2","2","yes","no","runner","no","","yes"
+"6","1810","blue","0","0","0","0","0","0","0","yes","no","0","0","0","0","0","0","0","yes","no","runner","no","","yes"
+"7","1811","blue","1","1","1","1","1","1","1","yes","no","1","1","1","1","1","1","1","yes","no","runner","no","","yes"
+"8","1812","blue","2","2","2","2","2","2","2","yes","no","2","2","2","2","2","2","2","yes","no","runner","no","","yes"
+"0","1810","blue","0","0","0","0","0","0","0","yes","no","0","0","0","0","0","0","0","yes","no","runner","no","","yes"
+"1","1811","blue","1","1","1","1","1","1","1","yes","no","1","1","1","1","1","1","1","yes","no","runner","no","","yes"
+"2","1812","blue","2","2","2","2","2","2","2","yes","no","2","2","2","2","2","2","2","yes","no","runner","no","","yes"
+"3","1810","blue","0","0","0","0","0","0","0","yes","no","0","0","0","0","0","0","0","yes","no","runner","no","","yes"
+"4","1811","blue","1","1","1","1","1","1","1","yes","no","1","1","1","1","1","1","1","yes","no","runner","no","","yes"
+"5","1812","blue","2","2","2","2","2","2","2","yes","no","2","2","2","2","2","2","2","yes","no","runner","no","","yes"
+"6","1810","blue","0","0","0","0","0","0","0","yes","no","0","0","0","0","0","0","0","yes","no","runner","no","","yes"
+"7","1811","blue","1","1","1","1","1","1","1","yes","no","1","1","1","1","1","1","1","yes","no","runner","no","","yes"
+"8","1812","blue","2","2","2","2","2","2","2","yes","no","2","2","2","2","2","2","2","yes","no","runner","no","","yes"
